backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pytest==8.3.3
openai>=1.51.0
python-multipart==0.0.20
python-dotenv==1.0.1
apscheduler==3.10.4
PyJWT==2.9.0
bcrypt>=4.0.0
pyyaml>=6.0
